+++ dove/src/lib.rs
@@ -23,3 +23,5 @@ pub mod manifest;
 pub mod stdout;
 /// for tests
 pub mod tests_helper;
+/// Build|Run transaction. Used in "dove run" and "dove tx"
+pub mod transaction;
